requirements.txt
python-dotenv
google-generativeai

langchain
langchain-community
langchain-google-genai

faiss-cpu
pypdf

streamlit

-e .